requirements.txt
fastapi
pynput
uvicorn
pyserial